add all tests to skilldialog tests
