+++ tests/Microsoft.Bot.Builder.Dialogs.Tests/SkillDialogTests.cs
@@ -8,12 +8,14 @@ using System.Globalization;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Adapters;
 using Microsoft.Bot.Builder.Skills;
 using Microsoft.Bot.Builder.Testing;
 using Microsoft.Bot.Connector;
 using Microsoft.Bot.Schema;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using Newtonsoft.Json.Linq;
 
 namespace Microsoft.Bot.Builder.Dialogs.Tests
 {
@@ -177,6 +179,149 @@ namespace Microsoft.Bot.Builder.Dialogs.Tests
             await Assert.ThrowsExceptionAsync<HttpRequestException>(async () => await client.SendActivityAsync<IMessageActivity>("irrelevant"));
         }
 
+        [TestMethod]
+        public async Task ShouldInterceptOAuthCardsForSSO()
+        {
+            string connectionName = "connectionName";
+            var firstResponse = new ExpectedReplies(new List<Activity>() { CreateOAuthCardAttachmentActivity("https://test") });
+            var mockSkillClient = new Mock<BotFrameworkClient>();
+            mockSkillClient
+                .SetupSequence(x => x.PostActivityAsync<ExpectedReplies>(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Uri>(), It.IsAny<Uri>(), It.IsAny<string>(), It.IsAny<Activity>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200, Body = firstResponse }))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200 }));
+
+            var conversationState = new ConversationState(new MemoryStorage());
+            var dialogOptions = CreateSkillDialogOptions(conversationState, mockSkillClient);
+
+            var sut = new SkillDialog(dialogOptions);
+            var activityToSend = CreateSendActivity();
+            var testAdapter = new TestAdapter(Connector.Channels.Test)
+            .Use(new AutoSaveStateMiddleware(conversationState));
+            var client = new DialogTestClient(testAdapter, sut, new BeginSkillDialogOptions { Activity = activityToSend, ConnectionName = connectionName }, conversationState: conversationState);
+            testAdapter.AddExchangeableToken(connectionName, Channels.Test, "user1", "https://test", "https://test1");
+            var finalActivity = await client.SendActivityAsync<IMessageActivity>("irrelevant");
+            Assert.IsNull(finalActivity);
+        }
+
+        [TestMethod]
+        public async Task ShouldNotInterceptOAuthCardsForEmptyConnectionName()
+        {
+            string connectionName = "connectionName";
+            var firstResponse = new ExpectedReplies(new List<Activity>() { CreateOAuthCardAttachmentActivity("https://test") });
+            var mockSkillClient = new Mock<BotFrameworkClient>();
+            mockSkillClient
+                .Setup(x => x.PostActivityAsync<ExpectedReplies>(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Uri>(), It.IsAny<Uri>(), It.IsAny<string>(), It.IsAny<Activity>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200, Body = firstResponse }));
+
+            var conversationState = new ConversationState(new MemoryStorage());
+            var dialogOptions = CreateSkillDialogOptions(conversationState, mockSkillClient);
+
+            var sut = new SkillDialog(dialogOptions);
+            var activityToSend = CreateSendActivity();
+            var testAdapter = new TestAdapter(Channels.Test)
+            .Use(new AutoSaveStateMiddleware(conversationState));
+            var client = new DialogTestClient(testAdapter, sut, new BeginSkillDialogOptions { Activity = activityToSend }, conversationState: conversationState);
+            testAdapter.AddExchangeableToken(connectionName, Channels.Test, "user1", "https://test", "https://test1");
+            var finalActivity = await client.SendActivityAsync<IMessageActivity>("irrelevant");
+            Assert.IsNotNull(finalActivity);
+            Assert.IsTrue(finalActivity.Attachments.Count == 1);
+        }
+
+        [TestMethod]
+        public async Task ShouldNotInterceptOAuthCardsForEmptyToken()
+        {
+            var firstResponse = new ExpectedReplies(new List<Activity>() { CreateOAuthCardAttachmentActivity("https://test") });
+            var mockSkillClient = new Mock<BotFrameworkClient>();
+            mockSkillClient
+                .Setup(x => x.PostActivityAsync<ExpectedReplies>(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Uri>(), It.IsAny<Uri>(), It.IsAny<string>(), It.IsAny<Activity>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200, Body = firstResponse }));
+
+            var conversationState = new ConversationState(new MemoryStorage());
+            var dialogOptions = CreateSkillDialogOptions(conversationState, mockSkillClient);
+
+            var sut = new SkillDialog(dialogOptions);
+            var activityToSend = CreateSendActivity();
+            var testAdapter = new TestAdapter(Channels.Test)
+            .Use(new AutoSaveStateMiddleware(conversationState));
+            var client = new DialogTestClient(testAdapter, sut, new BeginSkillDialogOptions { Activity = activityToSend }, conversationState: conversationState);
+
+            // dont add exchangeable token to test adapter
+            var finalActivity = await client.SendActivityAsync<IMessageActivity>("irrelevant");
+            Assert.IsNotNull(finalActivity);
+            Assert.IsTrue(finalActivity.Attachments.Count == 1);
+        }
+
+        [TestMethod]
+        public async Task ShouldNotInterceptOAuthCardsForTokenException()
+        {
+            string connectionName = "connectionName";
+            var firstResponse = new ExpectedReplies(new List<Activity>() { CreateOAuthCardAttachmentActivity("https://test") });
+            var mockSkillClient = new Mock<BotFrameworkClient>();
+            mockSkillClient
+                .Setup(x => x.PostActivityAsync<ExpectedReplies>(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Uri>(), It.IsAny<Uri>(), It.IsAny<string>(), It.IsAny<Activity>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200, Body = firstResponse }));
+
+            var conversationState = new ConversationState(new MemoryStorage());
+            var dialogOptions = CreateSkillDialogOptions(conversationState, mockSkillClient);
+
+            var sut = new SkillDialog(dialogOptions);
+            var activityToSend = CreateSendActivity();
+            var testAdapter = new TestAdapter(Channels.Test)
+            .Use(new AutoSaveStateMiddleware(conversationState));
+            var client = new DialogTestClient(testAdapter, sut, new BeginSkillDialogOptions { Activity = activityToSend, ConnectionName = connectionName }, conversationState: conversationState);
+            testAdapter.ThrowOnExchangeRequest(connectionName, Channels.Test, "user1", "https://test");
+            var finalActivity = await client.SendActivityAsync<IMessageActivity>("irrelevant");
+            Assert.IsNotNull(finalActivity);
+            Assert.IsTrue(finalActivity.Attachments.Count == 1);
+        }
+
+        [TestMethod]
+        public async Task ShouldNotInterceptOAuthCardsForBadRequest()
+        {
+            string connectionName = "connectionName";
+            var firstResponse = new ExpectedReplies(new List<Activity>() { CreateOAuthCardAttachmentActivity("https://test") });
+            var mockSkillClient = new Mock<BotFrameworkClient>();
+            mockSkillClient
+                .SetupSequence(x => x.PostActivityAsync<ExpectedReplies>(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Uri>(), It.IsAny<Uri>(), It.IsAny<string>(), It.IsAny<Activity>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 200, Body = firstResponse }))
+                .Returns(Task.FromResult(new InvokeResponse<ExpectedReplies> { Status = 409 }));
+
+            var conversationState = new ConversationState(new MemoryStorage());
+            var dialogOptions = SkillDialogTests.CreateSkillDialogOptions(conversationState, mockSkillClient);
+
+            var sut = new SkillDialog(dialogOptions);
+            var activityToSend = CreateSendActivity();
+            var testAdapter = new TestAdapter(Channels.Test)
+            .Use(new AutoSaveStateMiddleware(conversationState));
+            var client = new DialogTestClient(testAdapter, sut, new BeginSkillDialogOptions { Activity = activityToSend, ConnectionName = connectionName }, conversationState: conversationState);
+            testAdapter.AddExchangeableToken(connectionName, Channels.Test, "user1", "https://test", "https://test1");
+            var finalActivity = await client.SendActivityAsync<IMessageActivity>("irrelevant");
+            Assert.IsNotNull(finalActivity);
+            Assert.IsTrue(finalActivity.Attachments.Count == 1);
+        }
+
+        private static Activity CreateOAuthCardAttachmentActivity(string uri)
+        {
+            var oauthCard = new OAuthCard()
+            {
+                TokenExchangeResource = new TokenExchangeResource()
+                {
+                    Uri = uri
+                }
+            };
+            var attachment = new Attachment()
+            {
+                ContentType = OAuthCard.ContentType,
+                Content = JObject.FromObject(oauthCard)
+            };
+
+            var attachmentActivity = MessageFactory.Attachment(attachment);
+            attachmentActivity.Conversation = new ConversationAccount() { Id = Guid.NewGuid().ToString() };
+            attachmentActivity.From = new ChannelAccount("blah", "name");
+
+            return attachmentActivity as Activity;
+        }
+
         private static Mock<BotFrameworkClient> CreateMockSkillClient(Action<string, string, Uri, Uri, string, Activity, CancellationToken> captureAction, int returnStatus = 200)
         {
             var mockSkillClient = new Mock<BotFrameworkClient>();
@@ -196,6 +341,14 @@ namespace Microsoft.Bot.Builder.Dialogs.Tests
             }
 
             return mockSkillClient;
+        }
+
+        private Activity CreateSendActivity()
+        {
+            var activityToSend = (Activity)Activity.CreateMessageActivity();
+            activityToSend.DeliveryMode = DeliveryModes.ExpectReplies;
+            activityToSend.Text = Guid.NewGuid().ToString();
+            return activityToSend;
         }
 
         // Simple conversation ID factory for testing.
